turn get_supported_glsl_version into a free function
being able to call .get_supported_glsl_version() on a
glsl version was possible before, but is no longer.
